Re-organize and re-name a bit
