+++ spring-cloud-stream-metrics/src/main/java/org/springframework/cloud/stream/config/metrics/Emitter.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.stream.metrics;
+package org.springframework.cloud.stream.config.metrics;
 
 import org.springframework.cloud.stream.annotation.Output;
 import org.springframework.messaging.MessageChannel;
